refactor(utxo-lib): rename inputUpdate to fromHalfSigned
Issue: BG-57748
